chore: move loadConfig and saveConfig
