+++ common/index.js
@@ -1,3 +1,4 @@
 module.exports = {
   ...require('./cli-utils'),
+  ...require('./utils'),
 }
